Move `ConsoleUtils` to the `core` module
